+++ pkgs/google_generative_ai/test/generative_model_test.dart
@@ -69,6 +69,45 @@ void main() {
           ], null)));
     });
 
+    test('allows specifying a tuned model', () async {
+      final (client, model) = createModel('tunedModels/$defaultModelName');
+      final prompt = 'Some prompt';
+      final result = 'Some response';
+      client.stub(
+        Uri.parse('https://generativelanguage.googleapis.com/v1/'
+            'tunedModels/some-model:generateContent'),
+        {
+          'contents': [
+            {
+              'role': 'user',
+              'parts': [
+                {'text': prompt}
+              ]
+            }
+          ]
+        },
+        {
+          'candidates': [
+            {
+              'content': {
+                'role': 'model',
+                'parts': [
+                  {'text': result}
+                ]
+              }
+            }
+          ]
+        },
+      );
+      final response = await model.generateContent([Content.text(prompt)]);
+      expect(
+          response,
+          matchesGenerateContentResponse(GenerateContentResponse([
+            Candidate(
+                Content('model', [TextPart(result)]), null, null, null, null),
+          ], null)));
+    });
+
     group('generate unary content', () {
       test('can make successful request', () async {
         final (client, model) = createModel();
@@ -404,6 +443,56 @@ void main() {
             response,
             matchesEmbedContentResponse(
                 EmbedContentResponse(ContentEmbedding([0.1, 0.2, 0.3]))));
+      });
+    });
+
+    group('batch embed contents', () {
+      test('can make successful request', () async {
+        final (client, model) = createModel();
+        final prompt1 = 'Some prompt';
+        final prompt2 = 'Another prompt';
+        final embedding1 = [0.1, 0.2, 0.3];
+        final embedding2 = [0.4, 0.5, 1.6];
+        client.stub(
+          Uri.parse('https://generativelanguage.googleapis.com/v1/'
+              'models/some-model:batchEmbedContents'),
+          {
+            'requests': [
+              {
+                'content': {
+                  'role': 'user',
+                  'parts': [
+                    {'text': prompt1}
+                  ]
+                },
+                'model': 'models/$defaultModelName'
+              },
+              {
+                'content': {
+                  'role': 'user',
+                  'parts': [
+                    {'text': prompt2}
+                  ]
+                },
+                'model': 'models/$defaultModelName'
+              }
+            ]
+          },
+          {
+            'embeddings': [
+              {'values': embedding1},
+              {'values': embedding2}
+            ]
+          },
+        );
+        final response = await model.batchEmbedContents([
+          EmbedContentRequest(Content.text(prompt1)),
+          EmbedContentRequest(Content.text(prompt2))
+        ]);
+        expect(
+            response,
+            matchesBatchEmbedContentsResponse(BatchEmbedContentsResponse(
+                [ContentEmbedding(embedding1), ContentEmbedding(embedding2)])));
       });
     });
   });